// infrastructure/lib/open-api.ts
export const OpenAPI = {
  "openapi": "3.0.3",
  "info": {
    "title": "PSBAPP WtrPmp API",
    "version": "2021-05-27"
  },
  "paths": {
// ...
    "/search/location/text": {
      "get": {
        "tags": [
          "End User",
          "Search"
        ],
        "parameters": [
          {
            "name": "text",
            "in": "query",
            "description": "the text for location search",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "summary": "search text for coordinates",
        "description": "Searches the text and returns coordinates or suggestions",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "coordinates": {
                      "type": "object",
                      "properties": {
                        "latitude": {
                          "type": "string"
                        },
                        "longitude": {
                          "type": "string"
                        }
                      }
                    },
                    "suggestions": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
// ...
  },
// ...
}